Structurized front end components to folders (started to get lost in them)
