+++ GREEDY_WEB/ClientApp/components/All items page/FetchUserItems.tsx
@@ -2,8 +2,8 @@
 import axios from 'axios';
 import { Button, ButtonGroup, InputGroup, InputGroupAddon, Input, Form, FormGroup, Label, FormText } from 'reactstrap';
 import { ModalContainer, ModalDialog } from 'react-modal-dialog';
-import Constants from './Constants';
-import { EditItem } from './EditItem';
+import Constants from '../Shared/Constants';
+import { EditItem } from '../Photograph page/EditItem';
 
 interface Props {
     onRef: any
