+++ GREEDY_WEB/ClientApp/components/Photograph page/ImageUpload.tsx
@@ -1,8 +1,8 @@
 ﻿import * as React from 'react';
 import ImageUploader from 'react-images-upload';
 import axios from 'axios';
-import Constants from './Constants';
-import { Alert } from './Alert';
+import Constants from '../Shared/Constants';
+import { Alert } from '../Shared/Alert';
 
 interface Props {
     updateReceiptId: any;
