+++ GREEDY_WEB/ClientApp/components/Photograph page/PhotographPage.tsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import { RouteComponentProps } from 'react-router';
 import { ImageUpload } from './ImageUpload';
 import { FetchDataForUploadedReceipt } from './FetchDataForUploadedReceipt';
-import { Logo } from './Logo';
-import Constants from './Constants';
+import { Logo } from '../Shared/Logo';
+import Constants from '../Shared/Constants';
 
-export class Home extends React.Component<RouteComponentProps<{}>> {
+export class PhotographPage extends React.Component<RouteComponentProps<{}>> {
     child: any;
     constructor()
     {
