+++ GREEDY_WEB/GREEDY.csproj
@@ -48,22 +48,22 @@
   </ItemGroup>
 
   <ItemGroup>
-    <TypeScriptCompile Include="ClientApp\components\Alert.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\AllUserItems.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\Authorization.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\ChangeEmail.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\ChangePassword.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\Constants.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\EditItem.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\FetchUserItems.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\ItemPriceGraph.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\StatisticsPage.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\Logo.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\ForgotPassword.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\ImageUpload.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\UserLogin.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\UserRegistration.tsx" />
-    <TypeScriptCompile Include="ClientApp\components\UserSettings.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\Shared\Alert.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\All items page\AllUserItems.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\Shared\Authorization.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\User settings page\ChangeEmail.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\User settings page\ChangePassword.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\Shared\Constants.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\Photograph page\EditItem.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\All items page\FetchUserItems.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\Statistics page\ItemPriceGraph.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\Statistics page\StatisticsPage.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\Shared\Logo.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\User login\ForgotPassword.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\Photograph page\ImageUpload.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\User login\UserLogin.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\User login\UserRegistration.tsx" />
+    <TypeScriptCompile Include="ClientApp\components\User settings page\UserSettings.tsx" />
   </ItemGroup>
 
   <ItemGroup>
